박경호 : 22-12-20 기준 엔진 최신화 및 메모리릭/릴리즈빌드 체크 완료.

+++ GameEngineContents/OldCrowScreamBase.cpp
@@ -29,7 +29,7 @@ void OldCrowScreamBase::Update(float _DeltaTime)
 	float4 PlayerPos = Player::GetMainPlayer()->GetTransform().GetWorldPosition();
 	float4 MyPos = GetTransform().GetWorldPosition();
 	float Angle = DirToRot(PlayerPos, MyPos);
-	float RandY = GameEngineRandom::MainRandom.RandomInt(Angle - 30.f, Angle + 30.f);
+	float RandY = GameEngineRandom::MainRandom.RandomFloat(Angle - 30.f, Angle + 30.f);
